Split up styling and state
